setup VCR and refactored to add a forecast facade
