+++ spec/cassettes/google_city_coords.yml
@@ -1,0 +1,109 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: https://maps.googleapis.com/maps/api/geocode/json?address=atlanta%20%20ga&key=<GOOGLE_API_KEY>
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      User-Agent:
+      - Faraday v0.15.4
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Content-Type:
+      - application/json; charset=UTF-8
+      Date:
+      - Tue, 30 Jul 2019 22:49:17 GMT
+      Expires:
+      - Tue, 30 Jul 2019 22:51:17 GMT
+      Cache-Control:
+      - public, max-age=120
+      Vary:
+      - Accept-Language
+      Access-Control-Allow-Origin:
+      - "*"
+      Server:
+      - mafe
+      Content-Length:
+      - '444'
+      X-Xss-Protection:
+      - '0'
+      X-Frame-Options:
+      - SAMEORIGIN
+      Server-Timing:
+      - gfet4t7; dur=253
+      Alt-Svc:
+      - quic=":443"; ma=2592000; v="46,43,39"
+    body:
+      encoding: ASCII-8BIT
+      string: |
+        {
+           "results" : [
+              {
+                 "address_components" : [
+                    {
+                       "long_name" : "Atlanta",
+                       "short_name" : "Atlanta",
+                       "types" : [ "locality", "political" ]
+                    },
+                    {
+                       "long_name" : "Fulton County",
+                       "short_name" : "Fulton County",
+                       "types" : [ "administrative_area_level_2", "political" ]
+                    },
+                    {
+                       "long_name" : "Georgia",
+                       "short_name" : "GA",
+                       "types" : [ "administrative_area_level_1", "political" ]
+                    },
+                    {
+                       "long_name" : "United States",
+                       "short_name" : "US",
+                       "types" : [ "country", "political" ]
+                    }
+                 ],
+                 "formatted_address" : "Atlanta, GA, USA",
+                 "geometry" : {
+                    "bounds" : {
+                       "northeast" : {
+                          "lat" : 33.8876179,
+                          "lng" : -84.289389
+                       },
+                       "southwest" : {
+                          "lat" : 33.6478079,
+                          "lng" : -84.5518189
+                       }
+                    },
+                    "location" : {
+                       "lat" : 33.7489954,
+                       "lng" : -84.3879824
+                    },
+                    "location_type" : "APPROXIMATE",
+                    "viewport" : {
+                       "northeast" : {
+                          "lat" : 33.8876179,
+                          "lng" : -84.289389
+                       },
+                       "southwest" : {
+                          "lat" : 33.6478079,
+                          "lng" : -84.5518189
+                       }
+                    }
+                 },
+                 "place_id" : "ChIJjQmTaV0E9YgRC2MLmS_e_mY",
+                 "types" : [ "locality", "political" ]
+              }
+           ],
+           "status" : "OK"
+        }
+    http_version: 
+  recorded_at: Tue, 30 Jul 2019 22:49:17 GMT
+recorded_with: VCR 5.0.0
