+++ spec/services/google_geocode_service_spec.rb
@@ -2,15 +2,17 @@ require "rails_helper"
 
 describe 'google geocode service' do
   it 'can find coords of a city' do
-    location = 'atlanta, ga'
-    geocode = GoogleGeocodeService.new
+    VCR.use_cassette('google_city_coords') do
+      location = 'atlanta, ga'
+      geocode = GoogleGeocodeService.new
 
-    results = geocode.find_coords(location)
+      results = geocode.find_coords(location)
 
-    centered_location = results[:results][0][:geometry][:location]
-    expect(centered_location).to be_a Hash
-    expect(centered_location).to have_key(:lat)
-    expect(centered_location).to have_key(:lng)
+      centered_location = results[:results][0][:geometry][:location]
+      expect(centered_location).to be_a Hash
+      expect(centered_location).to have_key(:lat)
+      expect(centered_location).to have_key(:lng)
+    end
   end
 
   it 'can format an address' do
